- warnings fixes on 64 bit systems

+++ modules/tm/tm.c
@@ -438,7 +438,8 @@ static int script_init( struct sip_msg *foo, void *bar)
 
 static int mod_init(void)
 {
-	DBG( "TM - (size of cell=%d) initializing...\n",sizeof(struct cell));
+	DBG( "TM - (size of cell=%ld) initializing...\n", 
+			(long)sizeof(struct cell));
 	/* checking if we have sufficient bitmap capacity for given
 	   maximum number of  branches */
 	if (MAX_BRANCHES+1>31) {
